[SPARK-25521][SQL]Job id showing null in the logs when insert into command Job is finished.
## What changes were proposed in this pull request?
As part of  insert command  in FileFormatWriter, a job context is created for handling the write operation , While initializing the job context setupJob() API
in HadoopMapReduceCommitProtocol sets the jobid  in the Jobcontext configuration, Since we are directly getting the jobId from the map reduce JobContext the job id will come as
null in the logs. As a solution we shall display the job description uuid, user can get to know about their particular write job status with job description uuid.

## How was this patch tested?
Manually, verified the logs after the changes.